+++ CHANGELOG.md
@@ -1,10 +1,14 @@
-# Baseline Changelog
+# k8s-dogu-operator Changelog
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#2] Annotation `k8s-dogu-operator.cloudogu.com/ces-services` to Dogu-`Services` containing information of
+  related CES services. For more information see [Annotations](/docs/operations/annotations_en.md).
+
 ### Changed
 - [#11] **Breaking Change ahead!** The secret containing the dogu registry data was split and renamed from 
   `dogu-registry-com` to `k8s-dogu-operator-dogu-registry` and `k8s-dogu-operator-docker-registry`.
@@ -12,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   need to delete their old secret and create two new ones. The creation process is described 
   [here](docs/operations/configuring_the_docker_registry_en.md) and
   [here](docs/operations/configuring_the_dogu_registry_en.md).
+- [#2] Update makefiles to version 5.0.0
 
 ## [v0.2.0] - 2022-04-01
 ### Added